Added recording of observations in enjoy.py
